feat: add connectome augmentation (#100)
Add an augmentation step that serves to apply the connectome constraint
via a data augmentation loss.

The connectome specifies the presynaptic neurons that can drive changes
in voltage in post synaptic neurons. In each batch we add an
augmentation where the t->t+1 dynamics are constrained by this. Pick 100
random neurons in each batch, create a second synthetic batch where all
neurons that do not directly drive the 100 are set to zero. Add an
`aug_loss` component to reflect this.